# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
    patch:
      operationId: updateMyProfile
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
...
  /profile/security-settings:
    get:
      operationId: getMySecuritySettings
      tags: [profile]
      responses:
        "200":
          description: Security settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecuritySettings" }
    patch:
      operationId: updateMySecuritySettings
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecuritySettingsUpdate" }
      responses:
        "200":
          description: Updated security settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecuritySettings" }

  # ---------- Tenants ----------
...
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Log" }
    post:
      operationId: createLog
      tags: [logs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogInput" }
      responses:
        "201":
          description: Created log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Log" }
  /logs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateLog
      tags: [logs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogUpdate" }
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Log" }
    delete:
      operationId: deleteLog
      tags: [logs]
      responses:
        "204":
          description: Deleted

  # ---------- Metrics ----------
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Habit" }
    post:
      operationId: createHabit
      tags: [habits]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HabitInput" }
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Habit" }
  /habits/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateHabit
      tags: [habits]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HabitUpdate" }
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Habit" }
    delete:
      operationId: deleteHabit
      tags: [habits]
      responses:
        "204":
          description: Deleted
  /habits/{habitId}/checkins:
    parameters:
      - name: habitId
        in: path
        required: true
        schema: { type: integer }
    get:
      operationId: listHabitCheckins
      tags: [habits]
      responses:
        "200":
          description: List of checkins
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HabitCheckin" }
    post:
      operationId: createHabitCheckin
      tags: [habits]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HabitCheckinInput" }
      responses:
        "201":
          description: Created checkin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HabitCheckin" }

  # ---------- Projects & Tasks ----------
...
  /notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationUpdate" }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
    delete:
      operationId: deleteNotification
      tags: [notifications]
      responses:
        "204":
          description: Deleted
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
    patch:
      operationId: updateNotificationPreferences
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationPreferencesUpdate" }
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, fullName]
      properties:
        id: { type: string }
        fullName: { type: string }
        role: { type: ["string", "null"] }
        createdAt: { type: "string", format: date-time }
        updatedAt: { type: "string", format: date-time }
    ProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        role: { type: string }
...
    SecuritySettings:
      type: object
      required: [id, userId]
      properties:
        id: { type: string }
        userId: { type: string }
        mfaEnabled: { type: boolean }
        loginAttempts: { type: integer }
        lastFailedLogin: { type: ["string", "null"], format: date-time }
        accountLockedUntil: { type: ["string", "null"], format: date-time }
        passwordChangedAt: { type: ["string", "null"], format: date-time }
        sessionTimeoutMinutes: { type: integer }
    SecuritySettingsUpdate:
      type: object
      properties:
        mfaEnabled: { type: boolean }
        sessionTimeoutMinutes: { type: integer }

    # ---------- Tenants ----------
...
    Log:
      type: object
      required: [id, userId, logDate]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        hubId: { type: ["integer", "null"] }
        logDate: { type: string, format: date }
        metric: { type: ["string", "null"] }
        value: { type: ["number", "null"] }
        source: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LogInput:
      type: object
      required: [logDate]
      properties:
        tenantId: { type: string, format: uuid }
        hubId: { type: integer }
        logDate: { type: string, format: date }
        metric: { type: string }
        value: { type: number }
        source: { type: string }
        notes: { type: string }
    LogUpdate:
      type: object
      properties:
        metric: { type: string }
        value: { type: number }
        notes: { type: string }

    # ---------- Metrics ----------
...
    Habit:
      type: object
      required: [id, userId, name]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        name: { type: string }
        description: { type: ["string", "null"] }
        streak: { type: integer }
        lastCheckin: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    HabitInput:
      type: object
      required: [name]
      properties:
        tenantId: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
        description: { type: string }
    HabitUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        streak: { type: integer }
        lastCheckin: { type: string, format: date }
    HabitCheckin:
      type: object
      required: [id, habitId, date]
      properties:
        id: { type: integer }
        habitId: { type: integer }
        date: { type: string, format: date }
        done: { type: boolean }
        createdAt: { type: string, format: date-time }
    HabitCheckinInput:
      type: object
      required: [date]
      properties:
        date: { type: string, format: date }
        done: { type: boolean }

    # ---------- Projects / Tasks ----------
...
    Notification:
      type: object
      required: [id, userId, type, title, message, severity]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        severity: { type: string }
        metadata: {}
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        isRead: { type: boolean }
        readAt: { type: ["string", "null"], format: date-time }
        isResolved: { type: boolean }
        resolvedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    NotificationUpdate:
      type: object
      properties:
        isRead: { type: boolean }
        isResolved: { type: boolean }
    NotificationPreferences:
      type: object
      required: [id, userId]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        habitRemindersEnabled: { type: boolean }
        calendarAlertsEnabled: { type: boolean }
        performanceAlertsEnabled: { type: boolean }
        projectAlertsEnabled: { type: boolean }
        lifeEventAlertsEnabled: { type: boolean }
        weeklyReportsEnabled: { type: boolean }
        monthlyReportsEnabled: { type: boolean }
        intensityLevel: { type: ["string", "null"] }
        maxNotificationsPerHour: { type: ["integer", "null"] }
        quietHoursStart: { type: ["string", "null"] }
        quietHoursEnd: { type: ["string", "null"] }
    NotificationPreferencesUpdate:
      type: object
      properties:
        habitRemindersEnabled: { type: boolean }
        calendarAlertsEnabled: { type: boolean }
        performanceAlertsEnabled: { type: boolean }
        projectAlertsEnabled: { type: boolean }
        lifeEventAlertsEnabled: { type: boolean }
        weeklyReportsEnabled: { type: boolean }
        monthlyReportsEnabled: { type: boolean }
        intensityLevel: { type: string }
        maxNotificationsPerHour: { type: integer }
        quietHoursStart: { type: string }
        quietHoursEnd: { type: string }
...